refactor shortcodes to be more scalable - shortcodes are now a require instead of a long file, like `filers.js` this allows scalablity or extraction.
