+++ src/__flightdeck/shortcodes/copyright.js
@@ -1,0 +1,7 @@
+/**
+ * Get the current year - copyright
+ * @example {% year %}
+ */
+module.exports = (copyright) => {
+  return `&copy; ${new Date().getFullYear()} &nbsp;`;
+};
